break formatting out to a separate module
